# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AllocationResponse": {
        "additionalProperties": false,
        "properties": {
          "dimension": {
            "enum": [
              "asset_class",
              "currency",
              "issuer",
              "valuation_method",
              "raw_sector",
              "rating"
            ],
            "title": "Dimension",
            "type": "string"
          },
          "excluded_lot_count": {
            "title": "Excluded Lot Count",
            "type": "integer"
          },
          "excluded_value_kzt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Value Kzt"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AllocationItem"
            },
            "title": "Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "total_value_kzt": {
            "title": "Total Value Kzt",
            "type": "string"
          },
          "value_basis": {
            "enum": [
              "derived_carrying_value_kzt",
              "purchase_amount_kzt"
            ],
            "title": "Value Basis",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "dimension",
          "value_basis",
          "total_value_kzt",
          "items",
          "excluded_value_kzt",
          "excluded_lot_count"
        ],
        "title": "AllocationResponse",
        "type": "object"
      },
...
      "CalendarResponse": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/CalendarCounts"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CalendarEvent"
            },
            "title": "Items",
            "type": "array"
          },
          "report_date": {
            "title": "Report Date",
            "type": "string"
          },
          "settlement_total": {
            "$ref": "#/components/schemas/UnavailableSettlementTotal"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "report_date",
          "counts",
          "settlement_total",
          "items"
        ],
        "title": "CalendarResponse",
        "type": "object"
      },
...
      "CashResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CashBalance"
            },
            "title": "Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "items"
        ],
        "title": "CashResponse",
        "type": "object"
      },
...
      "DqAssignmentRequest": {
        "additionalProperties": false,
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "owner_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DqAssignmentRequest",
        "type": "object"
      },
      "DqAssignmentResponse": {
        "additionalProperties": false,
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_overdue": {
            "title": "Is Overdue",
            "type": "boolean"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "required": [
          "id",
          "owner_id",
          "due_date",
          "is_overdue"
        ],
        "title": "DqAssignmentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstrumentHoldingsResponse": {
        "additionalProperties": false,
        "properties": {
          "dividend_data_status": {
            "$ref": "#/components/schemas/DividendDataStatus"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstrumentHolding"
            },
            "title": "Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "value_basis": {
            "const": "derived_carrying_value_kzt",
            "title": "Value Basis",
            "type": "string"
          },
          "view": {
            "const": "instruments",
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "view",
          "value_basis",
          "items",
          "dividend_data_status"
        ],
        "title": "InstrumentHoldingsResponse",
        "type": "object"
      },
...
      "IssuesResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DataQualityIssue"
            },
            "title": "Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "items"
        ],
        "title": "IssuesResponse",
        "type": "object"
      },
...
      "LotHoldingsResponse": {
        "additionalProperties": false,
        "properties": {
          "dividend_data_status": {
            "$ref": "#/components/schemas/DividendDataStatus"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LotHolding"
            },
            "title": "Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "view": {
            "const": "lots",
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "view",
          "items",
          "dividend_data_status"
        ],
        "title": "LotHoldingsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OidcBearer": {
        "description": "Production OIDC access token. Development headers are local-only.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Portfolio Operations Insight API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/issues/{issue_id}/assign": {
      "post": {
        "operationId": "assignDqIssue",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DqAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DqAssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Assign or clear a data-quality finding's owner and due date",
        "tags": [
          "snapshots"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/allocations": {
      "get": {
        "operationId": "getSnapshotAllocations",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dimension",
            "required": false,
            "schema": {
              "default": "asset_class",
              "enum": [
                "asset_class",
                "currency",
                "issuer",
                "valuation_method",
                "raw_sector",
                "rating"
              ],
              "title": "Dimension",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "basis",
            "required": false,
            "schema": {
              "default": "derived_carrying",
              "enum": [
                "derived_carrying",
                "purchase"
              ],
              "title": "Basis",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read basis-aware portfolio allocation",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/calendar": {
      "get": {
        "operationId": "getSnapshotCalendar",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read operational dated events",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/cash": {
      "get": {
        "operationId": "getSnapshotCash",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read cash balances and source evidence",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/cash-calendar/export": {
      "get": {
        "description": "Download cash and the OSIP operational calendar in one workbook.",
        "operationId": "exportSnapshotCashCalendar",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Excel workbook with cash and operational calendar"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Export cash and operational calendar as an Excel workbook",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/holdings": {
      "get": {
        "operationId": "getSnapshotHoldings",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "lots",
              "enum": [
                "lots",
                "instruments"
              ],
              "title": "View",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "instruments": "#/components/schemas/InstrumentHoldingsResponse",
                      "lots": "#/components/schemas/LotHoldingsResponse"
                    },
                    "propertyName": "view"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/LotHoldingsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InstrumentHoldingsResponse"
                    }
                  ],
                  "title": "Response Getsnapshotholdings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read immutable lots or the instrument aggregation",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/holdings/export": {
      "get": {
        "description": "Download the ISIN-keyed instrument table and reconciled distributions.",
        "operationId": "exportSnapshotHoldings",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "basis",
            "required": false,
            "schema": {
              "default": "derived_carrying",
              "enum": [
                "derived_carrying",
                "purchase"
              ],
              "title": "Basis",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "term",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Term",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Class"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Excel workbook with the filtered instrument view"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Export the filtered instrument view as an Excel workbook",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/issues": {
      "get": {
        "operationId": "getSnapshotIssues",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read data-quality findings and acknowledgements",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/issues/export": {
      "get": {
        "description": "Download the exact filtered data-quality view for a published snapshot.",
        "operationId": "exportSnapshotIssues",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "term",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Term",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "blocker",
                "high",
                "medium",
                "low"
              ],
              "title": "Severity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Excel workbook with filtered data-quality findings"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Export filtered data-quality findings as an Excel workbook",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/lots/export": {
      "get": {
        "description": "Download the immutable source-lot detail for a published snapshot.",
        "operationId": "exportSnapshotLots",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Excel workbook with immutable source lots"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Export immutable source lots as an Excel workbook",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/overview": {
      "get": {
        "operationId": "getSnapshotOverview",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read snapshot totals with metric bases",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/provenance": {
      "get": {
        "description": "Return auditable lineage for every overview metric.\n\nSource metrics point to the exact workbook row and parsed field. Derived\nmetrics include their formula and the source inputs used by that formula.",
        "operationId": "getSnapshotProvenance",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotProvenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read exact source and derivation lineage for snapshot metrics",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/report-readiness": {
      "get": {
        "operationId": "getSnapshotReportReadiness",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read controlled export gates",
        "tags": [
          "snapshots"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/settlements": {
      "get": {
        "operationId": "getSnapshotSettlements",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OidcBearer": []
          }
        ],
        "summary": "Read raw-lineage-preserving settlements",
        "tags": [
          "snapshots"
        ]
      }
    },
...
  }
}